Move state and functions to custom hook file for TimerCountDown component
