+++ src/hooks/useTimerCountDownState.js
@@ -1,0 +1,47 @@
+import { useState } from 'react';
+import { playSound } from '../helper/countdown';
+
+const useTimerCountDownState = (task, isFocusing, toggleIsFocusing, editTask) => {
+  const [isPlaying, setIsPlaying] = useState(false);
+  const [key, setKey] = useState(0);
+  const handleStartClick = () => {
+    setIsPlaying(!isPlaying);
+  }
+  const resetTimer = () => {
+    setKey(key + 1);
+    setIsPlaying(false);
+  }
+  const manageCurrentTask = () => {
+    if (isFocusing) {
+      task.completedFocus += 1;
+    }
+    if (task.nbFocus === task.completedFocus) {
+      task.completed = true;
+    }
+    editTask(task);
+  }
+  const handleStopClick = () => {
+    resetTimer();
+  }
+  const handleSkipClick = () => {
+    toggleIsFocusing();
+    resetTimer();
+  }
+  const handleComplete = () => {
+    playSound();
+    manageCurrentTask();
+    toggleIsFocusing();
+    resetTimer();
+  }
+
+  return {
+    key,
+    isPlaying,
+    handleStartClick,
+    handleStopClick,
+    handleSkipClick,
+    handleComplete
+  }
+};
+
+export default useTimerCountDownState;
